Improved both r4 files, but I came across some problems with how the code was set up.
R5 IS THE BEST VERSION SO FAR!!!!

NetworkServer_r5: best implementation yet! Talks back and forth with the client. If the client requests the NEXT instruction set, it gives it to them. Also can issue the DONE command which stops all of the clients who are still searching. This version allows for more than one client to connect at one time.

NetworkServer_r5_withThreading: This version does not have all of the nice features of the r5, but it makes an attempt at using multiple threads. WILL RUN, but will throw an error near the end.

NetworkClient_r5: This is the best implementation yet! The client can connect to the server (yay). When the client is done with an instruction set, it can request for the server to give it the NEXT set. If it receives the DONE command from the server, the client stops its searching and halts.