Extracted prefix generation to get_prefix() and generalized it so the calling the get_prefix().next() method will yield a prefix compatible with either the crack_it or the get_chunk methods. Added get_chunk() method to generate an object of type Chunk to pass across the network. Added run_chunk() method to start cracking based on data from a passed Chunk object.     Some oddities in run_chunk() still, basically works, but not finished. Added tedt_chunking() method to implement tests of get_chunk and run_chunk. In order to allow for the use of a space as an input I changed from a space separating fields in my use of Chunk to newline characters '/n'.
Leaving this in latest stable despite the only semi-functional nature of the new methods because the methods previously available still work as they did before.

These two methods are preliminary versions and at the moment multiprocessing is not available. I am working out a good way to chunk out work to network clients. There are really two choices, send many chunks and each one is added to the queue on the client up to a total of, say, three times as many chunks as processors on the client. This would seem to be a good choice since it has some built in calibration based on machine capabilities. The other option would be to send much larger chunks, check the prefix over charsToCheck+1 which would be the equivalent of sending as many chunks as there are characters in the alphabet. This could be ten minutes or more for slower systems and/or large alphabets. On the other hand, it requires less code modification.